Fix bugs

+++ server/server_collections.go
@@ -19,7 +19,7 @@ func (server *Server) GetCollections(ctx context.Context, _ *emptypb.Empty) (*pb
 		response.Items = append(response.Items, &pb.Collections_Collection{
 			Id:        collection.Id.String(),
 			Name:      collection.Name,
-			Documents: collection.Documents,
+			Documents: uint32(collection.Documents),
 		})
 	}
